backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.1
PyMuPDF>=1.23.8
pikepdf>=8.10.1
pdfplumber>=0.10.3
pdf2docx>=0.5.6
Pillow>=10.1.0
pdf2image>=1.16.3
APScheduler>=3.10.4
python-magic-bin>=0.4.14
psutil>=5.9.6
python-docx>=1.1.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
rapidocr_onnxruntime>=1.3.24
